requirements0.txt
# DINOv3 轮胎 X 光异常检测 - 依赖包列表

# PyTorch 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# Hugging Face Transformers (用于加载 DINOv3)
transformers>=4.30.0

# 数据集处理
datasets>=2.14.0

# FAISS 向量检索（Meta 开源）
# GPU 版本：faiss-gpu
# CPU 版本：faiss-cpu
# 请根据实际环境选择其一
faiss-gpu>=1.7.2

# 科学计算
numpy>=1.21.0

# 图像处理
opencv-python>=4.5.0
Pillow>=9.0.0

# 机器学习评估工具
scikit-learn>=1.0.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 配置文件
pyyaml>=6.0

# 进度条
tqdm>=4.62.0

# 其他可选依赖
# pandas>=1.3.0  # 如果需要更详细的数据分析
# jupyter>=1.0.0  # 如果需要在 Jupyter 中运行
